+++ NEWS.md
@@ -1,5 +1,10 @@
 # Spectra 1.17
 
+## Change in 1.17.6
+
+- Accept labels argument as a `list` instead of a `character` in the 
+  plotting functions. 
+
 ## Change in 1.17.5
 
 - Move generics `processingChunkSize()`, `processingChunkFactor()` and